node/manager: Rename `fix_block` module to `check_blocks`
